+++ endpoint/opcua/src/test/java/de/fraunhofer/iosb/ilt/faaast/service/endpoint/opcua/OpcUaEndpointFullTest.java
@@ -45,6 +45,7 @@ import de.fraunhofer.iosb.ilt.faaast.service.endpoint.opcua.helper.assetconnecti
 import de.fraunhofer.iosb.ilt.faaast.service.endpoint.opcua.helper.assetconnection.TestOperationProviderConfig;
 import de.fraunhofer.iosb.ilt.faaast.service.model.messagebus.event.change.ElementCreateEventMessage;
 import de.fraunhofer.iosb.ilt.faaast.service.model.messagebus.event.change.ElementDeleteEventMessage;
+import de.fraunhofer.iosb.ilt.faaast.service.util.PortHelper;
 import java.io.IOException;
 import java.time.OffsetDateTime;
 import java.time.ZoneId;
@@ -90,17 +91,17 @@ public class OpcUaEndpointFullTest {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(OpcUaEndpointFullTest.class);
 
-    private static final int OPC_TCP_PORT = 18123;
     private static final long DEFAULT_TIMEOUT = 500;
-
-    private static final String ENDPOINT_URL = "opc.tcp://localhost:" + OPC_TCP_PORT;
 
     private static TestService service;
     private static int aasns;
+    private static int OPC_TCP_PORT;
+    private static String ENDPOINT_URL;
 
     @BeforeClass
     public static void startTest() throws Exception {
-
+        OPC_TCP_PORT = PortHelper.findFreePort();
+        ENDPOINT_URL = "opc.tcp://localhost:" + OPC_TCP_PORT;
         OpcUaEndpointConfig config = new OpcUaEndpointConfig.Builder()
                 .tcpPort(OPC_TCP_PORT)
                 .secondsTillShutdown(0)